blocktrack_backend/requirements.txt
asgiref==3.8.1
base58==2.1.1
certifi==2025.4.26
charset-normalizer==3.4.1
Django==5.2
djangorestframework==3.16.0
idna==3.10
multiaddr==0.0.9
netaddr==1.3.0
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
urllib3==2.4.0
varint==1.0.2
django-filter
psycopg2-binary
python-dotenv
kafka-python
django-cors-headers
drf_yasg